+++ components/posts/PostCard.tsx
@@ -3,7 +3,8 @@ import { COLORS } from 'styles/colors';
 import ProfileIcon from '@public/icons/profile.svg';
 import SadTomatoSvg from '@public/images/sad-tomato.svg';
 import { postDetailType } from './types';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
+import { getElapsedTimeStr, getIsInProgress } from './util';
 
 interface Props {
   _id: string;
@@ -17,36 +18,8 @@ export default function PostCard({ _id, participants, data, createdAt }: Props) 
 
   const [isInProgress, setIsInProgress] = useState(false);
 
-  const getIsInProgress = useCallback((startTime: string, endTime: string) => {
-    const startT = new Date(`${date} ${startTime}`).getTime();
-
-    let endT;
-    const endTimeHour = parseInt(endTime.split(':')[0]);
-    if (endTimeHour > 24) {
-      const newEndT = `${endTimeHour - 24}:${endTime.split(':')[1]}`; // 24시간 뺀 시간
-      const newDate = new Date(new Date(date).setDate(new Date(date).getDate() + 1)).toDateString(); // 하루 더한 날짜
-
-      endT = new Date(`${newDate} ${newEndT}`).getTime();
-    } else {
-      endT = new Date(`${date} ${endTime}`).getTime();
-    }
-    const nowT = Date.now();
-
-    if (nowT >= startT && nowT <= endT) setIsInProgress(true);
-  }, []);
-
-  const getElapsedTimeStr = useCallback((createdAt: string) => {
-    const dateObj = new Date(createdAt).getTime();
-    const elapsedTime = Date.now() - dateObj;
-    const elapsedHour = Math.floor(elapsedTime / 1000 / 60 / 60);
-    const elapsedDay = Math.floor(elapsedHour / 24);
-    if (elapsedHour > 24) return `${elapsedDay}일 전`;
-    else if (elapsedHour < 1) return `${Math.floor(elapsedTime / 1000 / 60)}분 전`;
-    else return `${elapsedHour}시간 전`;
-  }, []);
-
   useEffect(() => {
-    getIsInProgress(startTime, endTime);
+    setIsInProgress(getIsInProgress(date, startTime, endTime));
   }, [startTime, endTime]);
 
   return (
